Allows customizing where islands are created
NewIsland receives a new interface which determines the location
where the next island will be created

A default implementation is provided which mirrors the previous behavior
