+++ src/main/java/world/bentobox/bentobox/managers/island/NewIslandLocationStrategy.java
@@ -1,0 +1,13 @@
+package world.bentobox.bentobox.managers.island;
+
+import org.bukkit.Location;
+import org.bukkit.World;
+
+/**
+ * Determines the locations for new islands
+ * @author tastybento
+ *
+ */
+public interface NewIslandLocationStrategy {
+    public Location getNextLocation(World world);
+}
